requirements/base.txt
Django==4.2.13
django-ninja==1.2.0
djangorestframework==3.15.2
django-ckeditor==6.2.0
Pillow==9.0.1
better-profanity==0.7.0
django-celery==3.1.15
redis==5.0.7
django-ninja-jwt==5.3.2
psycopg2-binary==2.9.9
